progress

+++ lib/wanderer_kills/storage/killmail_store.ex
@@ -57,12 +57,12 @@ defmodule WandererKills.Storage.KillmailStore do
   """
   @impl true
   def init_tables! do
-    # Core tables
-    :ets.new(@killmails_table, [:set, :named_table, :public, {:read_concurrency, true}])
-    :ets.new(@system_killmails_table, [:set, :named_table, :public, {:read_concurrency, true}])
-    :ets.new(@system_kill_counts_table, [:set, :named_table, :public, {:read_concurrency, true}])
+    # Core tables - create only if they don't exist
+    ensure_table_exists(@killmails_table, [:set, :named_table, :public, {:read_concurrency, true}])
+    ensure_table_exists(@system_killmails_table, [:set, :named_table, :public, {:read_concurrency, true}])
+    ensure_table_exists(@system_kill_counts_table, [:set, :named_table, :public, {:read_concurrency, true}])
 
-    :ets.new(@system_fetch_timestamps_table, [
+    ensure_table_exists(@system_fetch_timestamps_table, [
       :set,
       :named_table,
       :public,
@@ -78,15 +78,15 @@ defmodule WandererKills.Storage.KillmailStore do
 
     # Event streaming tables (if enabled)
     if event_streaming_enabled?() do
-      :ets.new(@killmail_events_table, [
+      ensure_table_exists(@killmail_events_table, [
         :ordered_set,
         :named_table,
         :public,
         {:read_concurrency, true}
       ])
 
-      :ets.new(@client_offsets_table, [:set, :named_table, :public, {:read_concurrency, true}])
-      :ets.new(@counters_table, [:set, :named_table, :public, {:read_concurrency, true}])
+      ensure_table_exists(@client_offsets_table, [:set, :named_table, :public, {:read_concurrency, true}])
+      ensure_table_exists(@counters_table, [:set, :named_table, :public, {:read_concurrency, true}])
 
       # Initialize counters
       :ets.insert(@counters_table, {:event_counter, 0})
@@ -396,19 +396,37 @@ defmodule WandererKills.Storage.KillmailStore do
   """
   @impl true
   def clear do
-    :ets.delete_all_objects(@killmails_table)
-    :ets.delete_all_objects(@system_killmails_table)
-    :ets.delete_all_objects(@system_kill_counts_table)
-    :ets.delete_all_objects(@system_fetch_timestamps_table)
+    # Safely clear tables only if they exist
+    if :ets.info(@killmails_table) != :undefined do
+      :ets.delete_all_objects(@killmails_table)
+    end
+    
+    if :ets.info(@system_killmails_table) != :undefined do
+      :ets.delete_all_objects(@system_killmails_table)
+    end
+    
+    if :ets.info(@system_kill_counts_table) != :undefined do
+      :ets.delete_all_objects(@system_kill_counts_table)
+    end
+    
+    if :ets.info(@system_fetch_timestamps_table) != :undefined do
+      :ets.delete_all_objects(@system_fetch_timestamps_table)
+    end
 
     if event_streaming_enabled?() do
-      :ets.delete_all_objects(@killmail_events_table)
-      :ets.delete_all_objects(@client_offsets_table)
-      :ets.delete_all_objects(@counters_table)
-
-      # Reinitialize counters
-      :ets.insert(@counters_table, {:event_counter, 0})
-      :ets.insert(@counters_table, {:killmail_seq, 0})
+      if :ets.info(@killmail_events_table) != :undefined do
+        :ets.delete_all_objects(@killmail_events_table)
+      end
+      
+      if :ets.info(@client_offsets_table) != :undefined do
+        :ets.delete_all_objects(@client_offsets_table)
+      end
+      if :ets.info(@counters_table) != :undefined do
+        :ets.delete_all_objects(@counters_table)
+        # Reinitialize counters
+        :ets.insert(@counters_table, {:event_counter, 0})
+        :ets.insert(@counters_table, {:killmail_seq, 0})
+      end
     end
 
     :ok
@@ -609,5 +627,13 @@ defmodule WandererKills.Storage.KillmailStore do
     end
 
     :ok
+  end
+
+  # Helper function to ensure table exists before creation
+  defp ensure_table_exists(table_name, options) do
+    case :ets.info(table_name) do
+      :undefined -> :ets.new(table_name, options)
+      _ -> table_name
+    end
   end
 end
